Images merged with Places
